Move all listeners to SLCore.Listeners assembly (#5245)
Fixes #5244
