+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithStrongNames.txt
@@ -1,7 +1,7 @@
 ﻿---
 ################################
 # Assembly references report
-# Report date/time: 2024-02-19T15:44:01.6247761Z
+# Report date/time: 2024-02-23T09:41:14.2750375Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.11.0.223
@@ -447,11 +447,13 @@ Relative path: 'SonarLint.VisualStudio.SLCore.Listeners.dll'
 Referenced assemblies:
 - 'Microsoft.Alm.Authentication, Version=4.0.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
+- 'Newtonsoft.Json, Version=13.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed'
 - 'SonarLint.VisualStudio.ConnectedMode, Version=7.7.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
+- 'SonarLint.VisualStudio.Core, Version=7.7.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'SonarLint.VisualStudio.SLCore, Version=7.7.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.ComponentModel.Composition, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 6
+# Number of references: 8
 
 ---
 Assembly: 'SonarLint.VisualStudio.TypeScript, Version=7.7.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
