+++ src/SLCore.Listeners/Implementation/ProgressListener.cs
@@ -18,25 +18,32 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.Threading.Tasks;
 using SonarLint.VisualStudio.SLCore.Core;
 
-namespace SonarLint.VisualStudio.SLCore.Listener
+namespace SonarLint.VisualStudio.SLCore.Listeners.Implementation
 {
-    public class SelectProxiesResponse
-    {
-        public List<object> proxies = new List<object>();
-    }
-
     [Export(typeof(ISLCoreListener))]
     [PartCreationPolicy(CreationPolicy.Shared)]
-    public class ProxyConfigurationListener : ISLCoreListener
+    public class ProgressListener : ISLCoreListener
     {
-        public async Task<SelectProxiesResponse> SelectProxiesAsync(object parameters)
+        /// <summary>
+        /// Stub method for compability with SLCore. We do not support progress
+        /// </summary>
+        /// <param name="parameters">Parameter's here for compability we discard it</param>
+        public Task StartProgressAsync(object parameters)
         {
-            return new SelectProxiesResponse();
+            return Task.CompletedTask;
+        }
+
+        /// <summary>
+        /// Stub method for compability with SLCore. We do not support progress
+        /// </summary>
+        /// <param name="parameters">Parameter's here for compability we discard it</param>
+        public Task ReportProgressAsync(object parameters)
+        {
+            return Task.CompletedTask;
         }
     }
 }
