+++ src/SLCore.UnitTests/Integration/SLCoreIntegrationSmokeTest.cs
@@ -18,13 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using SonarLint.VisualStudio.Infrastructure.VS;
 using SonarLint.VisualStudio.SLCore.Common.Models;
 using SonarLint.VisualStudio.SLCore.Core;
-using SonarLint.VisualStudio.SLCore.Listener;
 using SonarLint.VisualStudio.SLCore.Service.Connection.Models;
 using SonarLint.VisualStudio.SLCore.Service.Lifecycle;
 using SonarLint.VisualStudio.SLCore.Service.Lifecycle.Models;
@@ -55,8 +52,8 @@ public class SLCoreIntegrationSmokeTest
 
         var slCoreServiceProvider = new SLCoreServiceProvider(new NoOpThreadHandler(), logger);
         slCoreServiceProvider.SetCurrentConnection(slCoreRunner.Rpc);
-        var slCoreListenerSetUp = new SLCoreListenerSetUp(new[] { new LoggerListener(new TestLogger(logToConsole: true)) });
-        slCoreListenerSetUp.Setup(slCoreRunner.Rpc);
+        //var slCoreListenerSetUp = new SLCoreListenerSetUp(new[] { new LoggerListener(new TestLogger(logToConsole: true)) });
+        //slCoreListenerSetUp.Setup(slCoreRunner.Rpc);
 
         slCoreServiceProvider.TryGetTransientService(out ISLCoreLifecycleService slCoreLifecycleService).Should()
             .BeTrue();
